Namespace the storage of notication configs against the flyweight directory
This will address an issue where removing a notification config in the
DSL does not cease notifications (until server restart). It also should
fix an issue where updating the auth details of a notification config
is not honored until a restart.

This happens because the notifier registration is persisted at the class
level in a static field. Because it isn't possible to detect the removal
of notifier configs from a parse means that notifier configs will not be
removed, and thus still respond to build stage notifications. Also,
because we keep notifier configs in a Set, any old notifier configs that
may not be replaced by updated configs if the only differences are in
the auth details (the set membership ignores auth).

So, to fix this, we namespace each config repo's notification config
registry by the flyweight directory and wipe the namespaced storage on
each parse and let the parsing rebuild the configs from scratch.

It should be noted that this does not cover 100% of the cases. If a user
changes the material config of the config repo itself, the flyweight
directory changes and we are unable to delete the old namespace because
the previous flyweight dirname is lost by the time we want to clear the
storage map. In an ideal world, we would namespace by config repo ID,
which does not change and add a hook to purge the namespace if the user
deletes the configrepo. However, the config repo ID is never passed in
the messages from GoCD to the plugin, so this cnanot be done without
changing the extension->plugin contract, which might be too much for
what it is worth to cover this edge case.